Add hypermedia API to replace XML-RPC and simple
Adds a new API that covers the usage of the XMP-RPC and the simple api.
https://github.com/pypa/warehouse/issues/284

This work is intended as a proof of concept for how a hypermedia API
could be implemented, setting up the patterns that can be extended to
cover the rest of the API. The new API introduces pagination to reduce
the load for list views.  Serializers are used to increase
maintainability and code reuse. Some filtering is added to meet the use
cases of XML-RPC. Many thanks to @werwty for hacking out an initial
implementation which has been squashed.

Introduces new dependencies:
   apispec==0.37.0 : Used to generate an api spec at /api/
   marshmallow==3.0.0b10 : Used to serialize responses
   PyYAML==3.12 : Dependency of apispec

All new endpoints are added to a new domain, "sandbox".

Note: Locally, all subdomains were treated just like the actual domain
so I was unable to make the subdomain works as expected. I followed the
pattern that forklift uses, and guessed how it should work.